move EnsureBufferUpdated and ChangeBuffer logic into a new BufferManager
